reorganize and rename rescue files
