+++ tools/rescue/src/cli_bootstrapper.rs
@@ -182,3 +182,64 @@ fn test_bootstrap_db() -> anyhow::Result<()> {
 
     Ok(())
 }
+
+pub fn check_rescue_bootstraps(db_path: &Path, blob_path: &Path) -> Result<()> {
+    let b = BootstrapOpts {
+        db_dir: db_path.to_owned(),
+        genesis_txn_file: blob_path.to_owned(),
+        waypoint_to_verify: None,
+        commit: false,
+        info: false,
+    };
+    if let Some(wp) = b.run()? {
+        println!("Rescue tx verified. Bootstrap with waypoint: {:?}", wp);
+    }
+    Ok(())
+}
+
+/// helper to apply a single step upgrade to a db
+/// usually for internal testing purposes (e.g. twin)
+pub fn one_step_apply_rescue_on_db(
+    db_to_change_path: &Path,
+    rescue_blob: &Path,
+) -> anyhow::Result<Waypoint> {
+    let mut waypoint: Option<Waypoint> = None;
+    let bootstrap = BootstrapOpts {
+        db_dir: db_to_change_path.to_owned(),
+        genesis_txn_file: rescue_blob.to_owned(),
+        waypoint_to_verify: None,
+        commit: false, // NOT APPLYING THE TX
+        info: false,
+    };
+
+    let waypoint_to_check = bootstrap.run()?.expect("could not get waypoint");
+
+    // give time for any IO to finish
+    std::thread::sleep(std::time::Duration::from_secs(1));
+
+    let bootstrap = BootstrapOpts {
+        db_dir: db_to_change_path.to_owned(),
+        genesis_txn_file: rescue_blob.to_owned(),
+        waypoint_to_verify: Some(waypoint_to_check),
+        commit: true, // APPLY THE TX
+        info: false,
+    };
+
+    let waypoint_post = bootstrap.run()?.expect("could not get waypoint");
+    assert!(
+        waypoint_to_check == waypoint_post,
+        "waypoints are not equal"
+    );
+    if let Some(w) = waypoint {
+        assert!(
+            waypoint_to_check == w,
+            "waypoints are not equal between nodes"
+        );
+    }
+    waypoint = Some(waypoint_to_check);
+    // }
+    match waypoint {
+        Some(w) => Ok(w),
+        None => anyhow::bail!("cannot generate consistent waypoint."),
+    }
+}
