+++ tools/rescue/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod cli_bootstrapper;
-pub mod one_step;
-pub mod replace_validators;
-pub mod rescue_cli;
-pub mod rescue_tx;
+pub mod cli_main;
 pub mod session_tools;
+pub mod transaction_factory;
