+++ tools/rescue/src/transaction_factory.rs
@@ -1,4 +1,4 @@
-use crate::{cli_bootstrapper::BootstrapOpts, session_tools};
+use crate::session_tools;
 use anyhow::Result;
 use diem_types::{
     account_address::AccountAddress,
@@ -9,12 +9,10 @@ use libra_framework::builder::framework_generate_upgrade_proposal::libra_compile
 use move_core_types::language_storage::CORE_CODE_ADDRESS;
 use std::path::{Path, PathBuf};
 
-pub fn save_rescue_blob(tx: Transaction, out_dir: &Path) -> Result<PathBuf> {
-    let file = out_dir.join("rescue.blob");
-
+pub fn save_rescue_blob(tx: Transaction, out_file: &Path) -> Result<PathBuf> {
     let bytes = bcs::to_bytes(&tx)?;
-    std::fs::write(&file, bytes.as_slice())?;
-    Ok(file)
+    std::fs::write(out_file, bytes.as_slice())?;
+    Ok(out_file.to_path_buf())
 }
 
 pub fn run_script_tx(script_path: &Path) -> Result<Transaction> {
@@ -59,18 +57,4 @@ pub fn register_vals(
         upgrade_mrb,
     )?;
     Ok(Transaction::GenesisTransaction(WriteSetPayload::Direct(cs)))
-}
-
-pub fn check_rescue_bootstraps(db_path: &Path, blob_path: &Path) -> Result<()> {
-    let b = BootstrapOpts {
-        db_dir: db_path.to_owned(),
-        genesis_txn_file: blob_path.to_owned(),
-        waypoint_to_verify: None,
-        commit: false,
-        info: false,
-    };
-    if let Some(wp) = b.run()? {
-        println!("Rescue tx verified. Bootstrap with waypoint: {:?}", wp);
-    }
-    Ok(())
 }
