+++ tools/rescue/tests/rescue_cli_creates_blob.rs
@@ -2,7 +2,7 @@ mod support;
 
 use libra_rescue::{
     cli_bootstrapper::BootstrapOpts,
-    rescue_cli::{RescueCli, Sub},
+    cli_main::{RescueCli, Sub},
 };
 use libra_smoke_tests::libra_smoke::LibraSmoke;
