+++ tools/testnet/src/lib.rs
@@ -3,4 +3,5 @@ pub mod cli_main;
 pub mod cli_swarm;
 pub mod config_twin;
 pub mod config_virgin;
+pub mod replace_validators_file;
 pub mod twin_swarm;
